+++ src/main/java/com/gmail/goosius/siegewar/listeners/SiegeWarSelfListener.java
@@ -3,7 +3,7 @@ package com.gmail.goosius.siegewar.listeners;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 
-import com.gmail.goosius.siegewar.SiegeController;
+import com.gmail.goosius.siegewar.utils.SiegeWarSiegeUtil;
 import com.gmail.goosius.siegewar.events.BattleSessionPreStartEvent;
 import com.gmail.goosius.siegewar.settings.SiegeWarSettings;
 import com.palmergames.bukkit.towny.object.Translatable;
@@ -13,7 +13,7 @@ public class SiegeWarSelfListener implements Listener {
 	@EventHandler
 	public void onBattleSessionPreStart(BattleSessionPreStartEvent event) {
 		if (SiegeWarSettings.cancelBattleSessionWhenNoActiveSieges()
-		&& (SiegeController.getSieges().isEmpty() || SiegeController.getSieges().stream().noneMatch(siege -> siege.getStatus().isActive()))) {
+		&& (SiegeWarSiegeUtil.getSieges().isEmpty() || SiegeWarSiegeUtil.getSieges().stream().noneMatch(siege -> siege.getStatus().isActive()))) {
 			event.setCancelled(true);
 			event.setCancellationMsg(Translatable.of("battle_session_cancelled_no_sieges").defaultLocale());
 		}
