+++ src/main/java/com/gmail/goosius/siegewar/playeractions/StartConquestSiege.java
@@ -1,7 +1,7 @@
 package com.gmail.goosius.siegewar.playeractions;
 
 
-import com.gmail.goosius.siegewar.SiegeController;
+import com.gmail.goosius.siegewar.utils.SiegeWarSiegeUtil;
 import com.gmail.goosius.siegewar.enums.SiegeType;
 import com.gmail.goosius.siegewar.enums.SiegeWarPermissionNodes;
 import com.gmail.goosius.siegewar.settings.SiegeWarSettings;
@@ -52,7 +52,7 @@ public class StartConquestSiege {
 		allowSiegeOrThrow(player, nationOfSiegeStarter, targetTown);
 
 		// Start a SiegeCamp that will kick off the Siege (or if SiegeAssemblies are disabled, start the Siege immediately.)
-		SiegeController.startSiegeCampProcess(player, bannerBlock, SiegeType.CONQUEST, targetTown, nationOfSiegeStarter, targetTown, townOfSiegeStarter, townBlock);
+		SiegeWarSiegeUtil.startSiegeCampProcess(player, bannerBlock, SiegeType.CONQUEST, targetTown, nationOfSiegeStarter, targetTown, townOfSiegeStarter, townBlock);
 	}
 
 	private static void allowSiegeOrThrow(Player player, Nation nationOfSiegeStarter, Town targetTown) throws TownyException {
